fix a few circular depdendencies
