+++ .changeset/dry-coins-attend.md
@@ -1,0 +1,5 @@
+---
+'@roadiehq/backstage-plugin-bugsnag': patch
+---
+
+Fix circular dependencies.
