+++ .changeset/silent-walls-deliver.md
@@ -1,0 +1,5 @@
+---
+'@roadiehq/backstage-plugin-argo-cd': patch
+---
+
+Fix circular dependencies.
