+++ plugins/frontend/backstage-plugin-argo-cd/src/index.ts
@@ -6,4 +6,4 @@ export {
   EntityArgoCDHistoryCard,
 } from './plugin';
 export * from './api';
-export { isArgocdAvailable } from './Router';
+export { isArgocdAvailable } from './conditions';
