+++ plugins/frontend/backstage-plugin-bugsnag/src/conditions.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Larder Software Limited
+ * Copyright 2022 Larder Software Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { Entity } from '@backstage/catalog-model';
+import { BUGSNAG_ANNOTATION } from './hooks/useBugsnagData';
 
-export * from './plugin';
-export * from './api';
-export { isBugsnagAvailable } from './conditions';
-export * from './components/ErrorsOverviewComponent';
-export * from './components/ErrorsTableComponent';
+export const isBugsnagAvailable = (entity: Entity) => {
+  return Boolean(entity?.metadata.annotations?.[BUGSNAG_ANNOTATION]);
+};
